Add functions to the js file
I added the major functions needed to play rock, paper, scissors. The
program can now prompt and save the player's response, randomly generate
a response from the computer and display whether the player won, lost or
has a tie.

The programs must later be modified to track the wins, losses and ties
of the player.